==> script/8_b_merge_appendix_1_2.py <==
from reportlab.lib.pagesizes import A4
from reportlab.platypus import SimpleDocTemplate, Paragraph, Image
from reportlab.lib.styles import getSampleStyleSheet
from reportlab.lib.units import inch, cm
from reportlab.platypus import PageBreak
import pandas as pd
from reportlab.lib import utils

# read data
nation_and_provinces_df = pd.read_excel('./data/nation_and_provinces.xlsx', sheet_name='Class')
fig1_data_df = pd.read_excel('./outcome/appendix/Figure Data/Fig.1 data.xlsx', sheet_name='panel A')

# empty pdf file
pdf_filename = './outcome/appendix/Supplementary Appendix 1_1.pdf'
doc = SimpleDocTemplate(pdf_filename, pagesize=A4, leftMargin=20, rightMargin=20, topMargin=15, bottomMargin=20)
story = []

# setting style
styles = getSampleStyleSheet()
styles['Title'].fontName = 'Times-Bold'
styles['Heading1'].fontName = 'Times-Bold'
styles['Heading2'].fontName = 'Times-Bold'
styles['Heading3'].fontName = 'Times-Bold'
styles['BodyText'].fontName = 'Times-Roman'
styles['Normal'].fontName = 'Times-Roman'
styles['Normal'].fontSize = 12
styles['Normal'].leading = styles['Heading1'].leading

# add space before title
story.append(Paragraph('<br/>' * 2, styles['Normal']))

# setting title
title_text = "Supplementary Appendix 1:"
title = Paragraph(title_text, styles['Title'])
story.append(title)

title_content = "Temporal shifts in 24 notifiable infectious diseases in China before and during the COVID-19 pandemic"
title = Paragraph(title_content, styles['Title'])
story.append(title)
story.append(PageBreak())

# insert image and add TOC entries
for index, row in fig1_data_df.iterrows():
    # if index == 2:
    #     break
    disease_name = row['disease']
    label = nation_and_provinces_df[nation_and_provinces_df['diseasename'] == disease_name]['label'].values[0]

    # add figure
    img_path = f'./outcome/appendix/Supplementary Appendix 1_1/{disease_name}.png'
    img = utils.ImageReader(img_path)
    image = Image(img_path, width=560, height=600)
    story.append(image)

    # add figure title
    title_text = f"Supplementary Fig. {index + 1}. Temporal variation in monthly incidence of {label} from January 2008 to December 2023 in China."
    title = Paragraph(title_text, styles['Heading2'])
    story.append(title)

    # add figure content
    title_content = f"""<b>(A)</b> The incidence of {label} in China from January 2008 to December 2023; 
    <b>(B)</b> The spatial distribution of cases in China; 
    <b>(C)</b> Temporal variation in monthly incidence among different provinces. 
    The heatmap represents the normalized monthly incidence data of each province, 
    and the color intensity corresponds to the normalized monthly incidence. 
    Provincial data in panel <b>(B)</b> and <b>(C)</b> before January 2020 sourced from the Chinese Public Health Science Data Center, 
    and data after January 2020 sourced from the provincial Notifiable Infectious Diseases Reports. 
    * Normalized monthly incidence > 10."""
    title = Paragraph(title_content, styles['Normal'])
    story.append(title)

    # add page break
    story.append(PageBreak())

doc.build(story)
